+++ Assets/01_Scripts/Cards/Effects/StatusEffect.cs
@@ -2,6 +2,6 @@
 public class StatusEffect
 {
     public BuffStatType statType;       //어떤 버프이고   
-    public int value;                   //얼마나 적용되고
+    public float value;                   //얼마나 적용되고
     public int duration;                //몇턴 적용되는지 표시
 }
